requirements.txt
Django
djangorestframework
asgiref
Pillow
numpy
pandas
beautifulsoup4
requests
aiohttp
